+++ pydash/__init__.py
@@ -430,6 +430,7 @@ from .utilities import (
     range_,
     range_right,
     result,
+    stub_array,
     stub_false,
     stub_string,
     stub_true,